versioningView: split into separate files
