+++ webapp/loggedin/admin/components/versioningView/index.js
@@ -1,0 +1,2 @@
+import VersioningView from './versioningView'
+export default VersioningView
